refactor(module): add new way of defining pestle module
This will allow us to have more control over adding modules definitions and to give the some default functionality like a declarative way of attaching events and comply with the basic signature that a module should have.

You could take a look at the new features in action in the example folder. Just run gulp to run the example.